Still trying to get rid of boost variant.  Still not working
I have an implementation of boost::any.  I can put stuff into it but
I can't get stuff out of it because I don't know the type.  I can't
just dynamic_cast because the data is pod, objects, pointers, vectors.
I need to hold onto the enum with the type like what they already have so
that I can get the type from the boost::any.  There is also a problem of
getting a constantInterpolate object which is derived from nemlobject.
Constant interpoloate is not in the moose interface code so we can't use it
and it needs to be saved as the nemlobject base class and treated as that.
So the template parameter for derived classes probably needs is_base_of arguments.